Misc: Add script for patching the query history
Adds `patch_query_history.py` in the `misc/scripts` directory. Its
function is to extend the existing VSCode query history with a new entry
whose JSON evaluator log summary points at a log that was created
outside of VSCode.

This enables the use of e.g. the Performance Comparison View on runs
that were not initiated from within VSCode.