Rust: Accept result regression.
Presumably because of a change in context, e.g. the module wrapping
this test now, or the different inputs.  Either way we *should* be
able to get the result, and what we *actually* get in our analysis
doesn't change here either.  We've just slightly changed the test
and exposed a gap.